:hammer: Fix data type :construction: Dijktra throws std::badalloc

+++ include/config.h
@@ -4,7 +4,7 @@
 namespace abm {
 namespace graph {
 //! Vertex id type
-using vertex_t = long;
+using vertex_t = long long;
 //! Weight type, that can be added with +
 using weight_t = double;
 }  // namespace graph
